Validations and CRUD operations

+++ backend/cardiagnostic-be/src/main/java/sbz/cardiagnosticbe/controller/RelatedFailuresController.java
@@ -6,6 +6,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import sbz.cardiagnosticbe.dto.failure.TRelatedFailures;
+import sbz.cardiagnosticbe.exception.RelatedFailureException;
+import sbz.cardiagnosticbe.repository.DetectedRelatedFailuresRepository;
+import sbz.cardiagnosticbe.service.DetectedRelatedFailuresService;
 import sbz.cardiagnosticbe.service.RelatedFailuresService;
 
 import javax.validation.Valid;
@@ -17,12 +20,18 @@ public class RelatedFailuresController {
     @Autowired
     private RelatedFailuresService relatedFailuresService;
 
+    @Autowired
+    private DetectedRelatedFailuresService detectedRelatedFailuresService;
+
     @RequestMapping(
             method = RequestMethod.POST,
             consumes = "application/json"
     )
     @PreAuthorize("hasAuthority('EXPERT')")
     public ResponseEntity addRelatedFailures(@Valid @RequestBody TRelatedFailures req) {
+        if (req.getFailuresIds().size() < 2)
+            throw new RelatedFailureException("Less than 2 failures selected");
+
         relatedFailuresService.add(req);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
@@ -33,7 +42,10 @@ public class RelatedFailuresController {
     )
     @PreAuthorize("hasAuthority('EXPERT')")
     public ResponseEntity removeRelatedFailuresEntity(@PathVariable Long id) {
-        relatedFailuresService.remove(id);
+        if (!detectedRelatedFailuresService.findByRelatedFailuresId(id).isEmpty()) {
+            relatedFailuresService.remove(id);
+        }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
